Inserção de validators nas models, uso do ToString na TokenDTO e validação antes de deletar em todas as controllers

+++ src/main/java/com/lead/catalagofilmes/controller/CategoriaController.java
@@ -1,5 +1,4 @@
 package com.lead.catalagofilmes.controller;
-
 import com.lead.catalagofilmes.models.Categoria;
 import com.lead.catalagofilmes.services.CategoriaService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,18 +7,16 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
 
-
 @Controller
 @RequestMapping(value = "/categoria")
 public class CategoriaController {
+
     @Autowired
     private CategoriaService categoriaService;
-
 
     @ResponseBody
     @GetMapping (value = "/All")
@@ -27,16 +24,16 @@ public class CategoriaController {
         try {
             List<Categoria> categorias = categoriaService.findAll();
             if (categorias.isEmpty()){
-                throw new Exception("Erro, sem categorias cadastradas");
+                return new ResponseEntity<String>("Erro, sem categorias cadastradas", HttpStatus.NOT_FOUND);
             }
-
-            return ResponseEntity.ok().body(categorias);
+            return new ResponseEntity<List<Categoria>>(categorias, HttpStatus.OK);
         } catch (Exception e){
             return new ResponseEntity<String>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
+
     @ResponseBody
-    @GetMapping(value = "/id{id}")
+    @GetMapping(value = "/id/{id}")
     public ResponseEntity<?> findById(@PathVariable Long id){
         Optional<Categoria> obj = categoriaService.findById(id);
         try {
@@ -53,11 +50,11 @@ public class CategoriaController {
     @Transactional
     public ResponseEntity<?> updateCategoria(@RequestBody @Validated Categoria obj){
         try {
-
-            if (categoriaService.verificaServiceCategoria(obj.getId())){
+            if (!categoriaService.verificaServiceCategoria(obj.getId())){
                 return new ResponseEntity<String>("Não foi encontrado a categoria especificada", HttpStatus.NOT_FOUND);
             }
-            return ResponseEntity.ok().body(categoriaService.update(obj));
+            Categoria atualizaCategoria = categoriaService.update(obj);
+            return ResponseEntity.ok().body(atualizaCategoria);
         } catch (Exception e){
             return new ResponseEntity<String>("Erro em atualizar a categoria",HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -66,8 +63,11 @@ public class CategoriaController {
     @DeleteMapping(value = "/categoriaDel/{id}")
     public ResponseEntity<?> delete(@PathVariable Long id){
         try {
+            if(categoriaService.verificaServiceCategoria(id)){
+                return new ResponseEntity<String>("Não foi encontrado a categoria especificada", HttpStatus.NOT_FOUND);
+            }
             categoriaService.deleteById(id);
-            return new ResponseEntity<String>("Categoria de id"+ id + " foi excluída com sucesso",HttpStatus.OK);
+            return new ResponseEntity<String>("Categoria de id "+ id + " foi excluída com sucesso",HttpStatus.OK);
         } catch (Exception e){
             return new ResponseEntity<String>("Erro ao deletar categoria", HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -76,7 +76,8 @@ public class CategoriaController {
     @PostMapping(value = "/create")
     public ResponseEntity<?> salvaCategoria(@RequestBody @Validated Categoria categoria){
         try {
-            return ResponseEntity.ok().body(categoriaService.save(categoria));
+            Categoria salvaCategoria = categoriaService.save(categoria);
+            return ResponseEntity.ok().body(salvaCategoria);
         }catch (Exception e){
             return new ResponseEntity<String>("Erro ao salvar categoria", HttpStatus.INTERNAL_SERVER_ERROR);
         }
