+++ src/main/java/com/lead/catalagofilmes/controller/IdiomaController.java
@@ -1,6 +1,4 @@
 package com.lead.catalagofilmes.controller;
-
-import com.lead.catalagofilmes.models.Categoria;
 import com.lead.catalagofilmes.models.Idioma;
 import com.lead.catalagofilmes.services.IdiomaService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,7 +7,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
@@ -37,29 +34,33 @@ public class IdiomaController {
     }
 
     @ResponseBody
-    @GetMapping(value = "/idioma{id}")
+    @GetMapping(value = "/idioma/{id}")
     public ResponseEntity<?> findByid(@PathVariable Long id){
         try {
             Optional<Idioma> obj = idiomaService.findById(id);
             if (!obj.isPresent()){
-                throw new Exception("Idioma não encontrado");
+                return new ResponseEntity<String>("Idioma não encontrado",HttpStatus.NOT_FOUND);
             }
             return ResponseEntity.ok().body(obj);
         }catch (Exception e){
-            return new ResponseEntity<>("Idioma não encontrado",HttpStatus.NOT_FOUND);
+            return new ResponseEntity<String>("Idioma não encontrado",HttpStatus.NOT_FOUND);
         }
     }
-    @DeleteMapping(value = "/delete{id}")
+
+    @DeleteMapping(value = "/delete/{id}")
     public ResponseEntity<?> deleteIdioma(@PathVariable Long id){
         try {
+            if(!idiomaService.verificaServiceIdioma(id)){
+                return new ResponseEntity<String>("Idioma não encontrado",HttpStatus.NOT_FOUND);
+            }
             idiomaService.deleteById(id);
             return new ResponseEntity<String>("Idioma de id "+ id +" foi excluído com sucesso", HttpStatus.OK);
         } catch (Exception e){
             return new ResponseEntity<String>("Erro ao deletar idioma",HttpStatus.INTERNAL_SERVER_ERROR);
 
         }
-
     }
+
     @PostMapping(value = "/create")
     public ResponseEntity<?> salvaCategoria(@RequestBody @Validated Idioma idioma){
         try {
@@ -73,14 +74,14 @@ public class IdiomaController {
     @Transactional
     public ResponseEntity<?> update(@RequestBody @Validated Idioma obj){
         try {
-            if (idiomaService.verificaServiceIdioma(obj.getId())){
+            if (!idiomaService.verificaServiceIdioma(obj.getId())){
                 return  new ResponseEntity<String>("Não foi encontrado o idioma especificado",HttpStatus.NOT_FOUND);
             }
-            return ResponseEntity.ok().body(idiomaService.update(obj));
+            Idioma salvaIdioma = idiomaService.update(obj);
+            return ResponseEntity.ok().body(salvaIdioma);
         } catch (Exception e){
             return new ResponseEntity<String>("Erro a atualizar o idioma", HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
-
 
 }
